+++ examples/apps/admin/src/plugins/index.js
@@ -4,6 +4,8 @@ import adminPlugins from "@webiny/app-admin/plugins";
 import i18nPlugins from "@webiny/app-i18n/admin/plugins";
 import securityPlugins from "@webiny/app-security/admin/plugins";
 import pageBuilderPlugins from "@webiny/app-page-builder/admin/plugins";
+import formBuilderPlugins from "@webiny/app-forms/admin/plugins";
+import formBuilderPageBuilderPlugins from "@webiny/app-forms/page-builder/admin/plugins";
 import cookiePolicyPlugins from "@webiny/app-cookie-policy/admin";
 import googleTagManagerPlugins from "@webiny/app-google-tag-manager/admin";
 import typeformPlugins from "@webiny/app-typeform/admin";
@@ -16,6 +18,8 @@ const plugins = [
     i18nPlugins,
     securityPlugins,
     pageBuilderPlugins,
+    formBuilderPlugins,
+    formBuilderPageBuilderPlugins,
     cookiePolicyPlugins,
     googleTagManagerPlugins,
     typeformPlugins,