+++ examples/apps/site/src/plugins/index.js
@@ -7,7 +7,7 @@ import mailchimpPlugins from "@webiny/app-mailchimp/render";
 import gtmPlugins from "@webiny/app-google-tag-manager/render";
 import i18nPlugins from "@webiny/app-i18n/site/plugins";
 import formsSitePlugins from "@webiny/app-forms/site/plugins";
-import formsCmsPlugins from "@webiny/app-forms/page-builder/site/plugins";
+import formsPbPlugins from "@webiny/app-forms/page-builder/site/plugins";
 
 const plugins = [
     fileUploadPlugin(),
@@ -19,7 +19,7 @@ const plugins = [
     gtmPlugins,
     i18nPlugins,
     formsSitePlugins,
-    formsCmsPlugins
+    formsPbPlugins
 ];
 
 export default plugins;