+++ packages/api-forms/src/plugins/graphql/formResolvers/utils/getListPublishedFormsResolver.js
@@ -1,0 +1,9 @@
+// @flow
+export default (context: Object) => {
+    const plugin = context.plugins.byName("forms-resolver-list-published-forms");
+    if (!plugin) {
+        throw Error(`Resolver plugin "forms-resolver-list-published-forms" is not configured!`);
+    }
+
+    return plugin;
+};