+++ packages/app-forms/src/admin/components/FormEditor/DragPreview.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { useDragLayer } from "react-dnd";
 
 const layerStyles = {
@@ -27,6 +27,7 @@ const onOffsetChange = monitor => () => {
 };
 
 export default () => {
+    const [dragHelperOpacity, setDragHelperOpacity] = useState(0);
     const { isDragging } = useDragLayer(monitor => {
         if (!subscribedToOffsetChange) {
             monitor.subscribeToOffsetChange(onOffsetChange(monitor));
@@ -38,13 +39,44 @@ export default () => {
         };
     });
 
+    useEffect(() => {
+        return () => {
+            subscribedToOffsetChange = false;
+            dragPreviewRef = null;
+        };
+    }, []);
+
+    // We track the value of "isDragging" and apply opacity=1 (after 100ms), when it switches to true.
+    // Without this, the drag cursor would be shown in the top-left corner for a short amount of time, and then it
+    // would be repositioned correctly. Definitely looks like a glitch. This also adds a nice little fade-in effect.
+    useEffect(() => {
+        if (isDragging) {
+            setTimeout(() => {
+                setDragHelperOpacity(isDragging ? 1 : 0);
+            }, 100);
+        } else {
+            setDragHelperOpacity(0);
+        }
+    }, [isDragging]);
+
+    if (!isDragging) {
+        return null;
+    }
+
     if (!isDragging) {
         return null;
     }
 
     return (
         <div style={layerStyles}>
-            <div style={{ display: "block" }} ref={el => (dragPreviewRef = el)}>
+            <div
+                ref={el => (dragPreviewRef = el)}
+                style={{
+                    transition: "opacity .25s ease-in-out",
+                    display: "block",
+                    opacity: dragHelperOpacity
+                }}
+            >
                 <div
                     style={{
                         width: 30,