+++ packages/app-forms/src/page-builder/admin/plugins/components/FormElement.js
@@ -25,14 +25,14 @@ const FormElement = (props: Object) => {
     if (form.revision) {
         const props = {
             preview: true,
-            parent: undefined,
-            revision: undefined
+            parentId: undefined,
+            revisionId: undefined
         };
 
         if (form.revision === "latest") {
-            props.parent = form.parent;
+            props.parentId = form.parent;
         } else {
-            props.revision = form.revision;
+            props.revisionId = form.revision;
         }
 
         render = <FormsForm {...props} />;