+++ packages/cli/create/template/apps/site/src/plugins/index.js
@@ -6,6 +6,8 @@ import typeformPlugins from "@webiny/app-typeform/render";
 import mailchimpPlugins from "@webiny/app-mailchimp/render";
 import gtmPlugins from "@webiny/app-google-tag-manager/render";
 import i18nPlugins from "@webiny/app-i18n/site/plugins";
+import formsSitePlugins from "@webiny/app-forms/site/plugins";
+import formsPbPlugins from "@webiny/app-forms/page-builder/site/plugins";
 
 const plugins = [
     fileUploadPlugin(),
@@ -15,7 +17,9 @@ const plugins = [
     typeformPlugins,
     mailchimpPlugins,
     gtmPlugins,
-    i18nPlugins
+    i18nPlugins,
+    formsSitePlugins,
+    formsPbPlugins
 ];
 
 export default plugins;