+++ components/serverless-files/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-files@2.3.2-next.0...@webiny/serverless-files@2.3.2) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/serverless-files
+
+
+
+
+
+## [2.3.2-next.0](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-files@2.3.1...@webiny/serverless-files@2.3.2-next.0) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/serverless-files
+
+
+
+
+
 ## [2.3.1](https://github.com/webiny/webiny-js/compare/@webiny/serverless-files@2.3.0...@webiny/serverless-files@2.3.1) (2019-11-21)