+++ packages/api-cookie-policy/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/webiny/webiny-js/compare/@webiny/api-cookie-policy@2.1.1-next.0...@webiny/api-cookie-policy@2.1.1) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/api-cookie-policy
+
+
+
+
+
+## [2.1.1-next.0](https://github.com/webiny/webiny-js/compare/@webiny/api-cookie-policy@2.1.0...@webiny/api-cookie-policy@2.1.1-next.0) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/api-cookie-policy
+
+
+
+
+
 # [2.1.0](https://github.com/webiny/webiny-js/compare/@webiny/api-cookie-policy@2.0.1...@webiny/api-cookie-policy@2.1.0) (2019-11-20)