+++ packages/api-plugin-form-builder-resolvers-mongodb/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/webiny/webiny-js/compare/@webiny/api-plugin-form-builder-resolvers-mongodb@1.2.0-next.1...@webiny/api-plugin-form-builder-resolvers-mongodb@1.2.0) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/api-plugin-form-builder-resolvers-mongodb
+
+
+
+
+
+# [1.2.0-next.1](https://github.com/webiny/webiny-js/compare/@webiny/api-plugin-form-builder-resolvers-mongodb@1.2.0-next.0...@webiny/api-plugin-form-builder-resolvers-mongodb@1.2.0-next.1) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/api-plugin-form-builder-resolvers-mongodb
+
+
+
+
+
+# 1.2.0-next.0 (2019-12-04)
+
+
+### Bug Fixes
+
+* remove "private" flag ([87982ab](https://github.com/webiny/webiny-js/commit/87982ab6858985b0a791c288fa86f29b117a1f5b))
+
+
+### Features
+
+* form builder mongodb resolvers ([f23d436](https://github.com/webiny/webiny-js/commit/f23d436a53d7e09f6849ad4c7759964d9958f87d))
+
+
+
+
+
 ## [1.1.1](https://github.com/webiny/webiny-js/compare/@webiny/api-plugin-forms-resolvers-mongodb@1.1.0...@webiny/api-plugin-forms-resolvers-mongodb@1.1.1) (2019-11-21)
 
 **Note:** Version bump only for package @webiny/api-plugin-forms-resolvers-mongodb