+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1](https://github.com/webiny/webiny-js/compare/@webiny/ui@2.0.1-next.1...@webiny/ui@2.0.1) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/ui
+
+
+
+
+
+## [2.0.1-next.1](https://github.com/webiny/webiny-js/compare/@webiny/ui@2.0.1-next.0...@webiny/ui@2.0.1-next.1) (2019-12-04)
+
+**Note:** Version bump only for package @webiny/ui
+
+
+
+
+
+## [2.0.1-next.0](https://github.com/webiny/webiny-js/compare/@webiny/ui@2.0.0...@webiny/ui@2.0.1-next.0) (2019-12-04)
+
+
+### Bug Fixes
+
+* page builder settings sliders ([#631](https://github.com/webiny/webiny-js/issues/631)) ([fb3b52d](https://github.com/webiny/webiny-js/commit/fb3b52d20105b1ae34dcdc7554057c75cc747cc4))
+
+
+
+
+
 # [2.0.0](https://github.com/webiny/webiny-js/compare/@webiny/ui@0.1.12...@webiny/ui@2.0.0) (2019-10-29)
 
 **Note:** Version bump only for package @webiny/ui